+++ app/src/main/java/com/alorma/notifix/domain/repository/TriggersRepository.kt
@@ -9,11 +9,11 @@ import javax.inject.Inject
 
 class TriggersRepository @Inject constructor(private val cacheTriggersDataSource: CacheTriggersDataSource) {
 
-    fun getById(id: Long): Single<NotificationTrigger> =
+    fun getById(id: Int): Single<NotificationTrigger> =
             cacheTriggersDataSource.get(id).subscribeOnIO()
 
     fun getByPayload(payloadLauncher: PayloadLauncher): Single<NotificationTrigger> =
             cacheTriggersDataSource.get(payloadLauncher).subscribeOnIO()
 
-    fun createTrigger(trigger: NotificationTrigger): Single<Long> = cacheTriggersDataSource.save(trigger).subscribeOnIO()
+    fun createTrigger(trigger: NotificationTrigger): Single<Int> = cacheTriggersDataSource.save(trigger).subscribeOnIO()
 }
